+++ craftr/lib/craftr.ext.platform.py
@@ -18,11 +18,11 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-''' This module imports a platform implementation based on
-:data:`sys.platform` and exports all its contents into the global
-module scope.
+''' This module represents the current platform that Craftr is running
+on by importing the correct implementation based on :data:`sys.platform`.
 
-:Submodules:
+Available Implementations
+-------------------------
 
 .. toctree::
   :maxdepth: 2
@@ -31,25 +31,27 @@ module scope.
   platform_darwin
   platform_linux
   platform_win32
+
+Contents
+--------
 '''
 
-__all__ = ['cc', 'cxx', 'ld', 'ar']
+__all__ = ['WIN32', 'DARWIN', 'LINUX', 'CYGWIN', 'cc', 'cxx', 'ld', 'ar']
 
 from craftr import import_module
 from craftr.magic import Proxy
 import sys
 
-# Identifiers for the current platform names.
-WIN32 = 'win'
-DARWIN = 'mac'
-LINUX = 'linux'
-CYGWIN = 'cygwin'
+WIN32 = 'win'     #: Windows platform name
+DARWIN = 'mac'    #: Mac OS platform name
+LINUX = 'linux'   #: Linux platform name
+CYGWIN = 'cygwin' #: Cygwin platform name
 
 # Wildcard-import the module for the current platform.
 import_module(__name__ + '.' + sys.platform, globals(), '*')
 
 # get_tool() imported from the current platform module.
-cc = Proxy(get_tool, 'cc')
-cxx = Proxy(get_tool, 'cxx')
-ld = Proxy(get_tool, 'ld')
-ar = Proxy(get_tool, 'ar')
+cc = Proxy(get_tool, 'cc')    #: The C compiler retrieved with :func:`platform.get_tool`
+cxx = Proxy(get_tool, 'cxx')  #: The C++ compiler retrieved with :func:`platform.get_tool`
+ld = Proxy(get_tool, 'ld')    #: The linker retrieved with :func:`platform.get_tool`
+ar = Proxy(get_tool, 'ar')    #: The archiver retrieved with :func:`platform.get_tool`